// web/lib/api/generated/model/gatewayModeEnum.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Inventory Platform API
 * Multi-tenant B2B inventory & ordering platform. All money values are decimal strings.
 * OpenAPI spec version: 1.0.0
 */

/**
 * * `TEST` - Test
 * * `LIVE` - Live
 */
export type GatewayModeEnum = typeof GatewayModeEnum[keyof typeof GatewayModeEnum];


export const GatewayModeEnum = {
  TEST: 'TEST',
  LIVE: 'LIVE',
} as const;