// web/lib/api/generated/model/gatewayProviderEnum.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Inventory Platform API
 * Multi-tenant B2B inventory & ordering platform. All money values are decimal strings.
 * OpenAPI spec version: 1.0.0
 */

/**
 * * `RAZORPAY` - Razorpay
 * * `MOCK` - Test gateway (development)
 */
export type GatewayProviderEnum = typeof GatewayProviderEnum[keyof typeof GatewayProviderEnum];


export const GatewayProviderEnum = {
  RAZORPAY: 'RAZORPAY',
  MOCK: 'MOCK',
} as const;